Refactor : 회원가입, 로그인 폴더명 수정
